Remove tap_machine object and roll everything into the tap fitting method.
