+++ paysage/fit.py
@@ -413,7 +413,8 @@ def persistent_contrastive_divergence(vdata, model, sampler, steps=1):
 # alias
 pcd = persistent_contrastive_divergence
 
-def tap(vdata, model, sampler, positive_steps=1):
+def tap(vdata, model, sampler, positive_steps=1, terms=2, init_lr_EMF=0.1, tolerance_EMF=1e-4,
+        max_iters_EMF=25, num_random_samples=1):
     """
     Compute the gradient using the Thouless-Anderson-Palmer (TAP)
     mean field approximation.
@@ -429,16 +430,29 @@ def tap(vdata, model, sampler, positive_steps=1):
         sampler: for marginal free energy
         positive_steps: steps to sample MCMC for positive phase
 
+        TAP free energy computation parameters:
+            init_lr float: initial learning rate which is halved whenever necessary to enforce descent.
+            tol float: tolerance for quitting minimization.
+            max_iters: maximum gradient decsent steps
+            num_random_samples: number of Gibbs FE seeds to start from random
+            num_persistent_samples: number of persistent magnetization parameters to keep as seeds
+                for Gibbs FE estimation.
+
     Returns:
-        gradient
+        gradient object
 
     """
-    if model.num_layers != 2:
-        raise ValueError("Currently only 2-layer models are supported for the tap method")
+    if num_random_samples <= 0:
+        raise ValueError("Must specify at least one random or persistent sample for Gibbs FE seeding")
+    if terms not in [1, 2, 3]:
+        raise ValueError("Must specify one, two, or three terms in TAP expansion training method")
+
     data_state = State.from_visible(vdata, model)
     sampler.set_positive_state(data_state)
     sampler.update_positive_state(positive_steps)
-    return model.TAP_gradient(data_state, None)
+
+    return model.TAP_gradient(data_state, None, num_random_samples, 0, [],
+                              init_lr_EMF, tolerance_EMF, max_iters_EMF, terms)
 
 
 class StochasticGradientDescent(object):
